python -m submodules.release_tool.release make-links  

Before:
1.29s user 0.29s system 1% cpu 1:43.55 total

After:
1.33s user 0.31s system 3% cpu 46.520 total